+++ src/base/AbstractFieldProcessor.php
@@ -6,7 +6,12 @@ use craft\base\Element;
 use craft\base\Field;
 use craft\base\FieldInterface;
 use RuntimeException;
-use venveo\bulkedit\services\BulkEdit;
+use venveo\bulkedit\fields\strategies\Add;
+use venveo\bulkedit\fields\strategies\Divide;
+use venveo\bulkedit\fields\strategies\Merge;
+use venveo\bulkedit\fields\strategies\Multiply;
+use venveo\bulkedit\fields\strategies\Replace;
+use venveo\bulkedit\fields\strategies\Subtract;
 
 abstract class AbstractFieldProcessor implements FieldProcessorInterface
 {
@@ -24,32 +29,21 @@ abstract class AbstractFieldProcessor implements FieldProcessorInterface
     /**
      * @return array
      */
-    public static function getSupportedNativeFields(): array {
+    public static function getSupportedNativeFields(): array
+    {
         return [];
     }
 
-    public static function processElementField(Element $element, Field $field, $strategy, $newValue)
+    public static function processElementField(Element $element, Field $field, string $strategy, mixed $newValue)
     {
-        switch ($strategy) {
-            case BulkEdit::STRATEGY_REPLACE:
-                static::performReplacement($element, $field, $newValue);
-                break;
-            case BulkEdit::STRATEGY_MERGE:
-                static::performMerge($element, $field, $newValue);
-                break;
-            case BulkEdit::STRATEGY_SUBTRACT:
-                static::performSubtraction($element, $field, $newValue);
-                break;
-            case BulkEdit::STRATEGY_ADD:
-                static::performAddition($element, $field, $newValue);
-                break;
-            case BulkEdit::STRATEGY_MULTIPLY:
-                static::performMultiplication($element, $field, $newValue);
-                break;
-            case BulkEdit::STRATEGY_DIVIDE:
-                static::performDivision($element, $field, $newValue);
-                break;
-        }
+        match ($strategy) {
+            Replace::class => static::performReplacement($element, $field, $newValue),
+            Merge::class => static::performMerge($element, $field, $newValue),
+            Subtract::class => static::performSubtraction($element, $field, $newValue),
+            Add::class => static::performAddition($element, $field, $newValue),
+            Multiply::class => static::performMultiplication($element, $field, $newValue),
+            Divide::class => static::performDivision($element, $field, $newValue),
+        };
     }
 
     public static function performReplacement(Element $element, Field $field, $value)