+++ src/base/FieldStrategyInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace venveo\bulkedit\base;
+
+interface FieldStrategyInterface {
+    public static function displayName(): string;
+}