+++ src/queue/jobs/SaveBulkEditJob.php
@@ -14,10 +14,8 @@ use Craft;
 use craft\base\ElementInterface;
 use craft\queue\BaseJob;
 use Throwable;
+use venveo\bulkedit\models\EditContext;
 use venveo\bulkedit\Plugin;
-use venveo\bulkedit\records\EditContext;
-use venveo\bulkedit\records\History;
-use yii\base\Exception;
 
 /**
  * @author    Venveo
@@ -26,15 +24,7 @@ use yii\base\Exception;
  */
 class SaveBulkEditJob extends BaseJob
 {
-    // Public Properties
-    // =========================================================================
-
-
-    /** @var EditContext */
-    public $context;
-
-    // Public Methods
-    // =========================================================================
+    public EditContext $context;
 
     /**
      * Saves bulk edited elements
@@ -43,15 +33,10 @@ class SaveBulkEditJob extends BaseJob
      */
     public function execute($queue = null): void
     {
-        $elementHistories = Plugin::$plugin->bulkEdit->getPendingElementsHistoriesFromContext($this->context);
-        $totalSteps = $elementHistories->count();
+        $totalSteps = $this->context->total;
         try {
             $currentRow = 0;
-            /**
-             * @var History $elementHistory
-             */
-            foreach ($elementHistories->each() as $elementHistory) {
-                $elementId = $elementHistory->elementId;
+            foreach ($this->context->elementIds as $elementId) {
                 /** @var ElementInterface $element */
                 $element = Craft::$app->getElements()->getElementById($elementId, null, $this->context->siteId);
                 if (!$element) {
@@ -59,14 +44,8 @@ class SaveBulkEditJob extends BaseJob
                     continue;
                 }
 
-                if ($element::class !== $this->context->elementType) {
-                    throw new \Exception('Unexpected element type encountered!');
-                }
-
-                $history = Plugin::$plugin->bulkEdit->getPendingHistoryFromContext($this->context, $element->id)->all();
                 try {
-                    Craft::info('Starting processing bulk edit job', __METHOD__);
-                    Plugin::$plugin->bulkEdit->processHistoryItemsForElement($history, $element);
+                    Plugin::$plugin->bulkEdit->processElementWithContext($element, $this->context);
                 } catch (\Exception $exception) {
                     Craft::error('Could not save element in bulk edit job... ' . $exception->getMessage(), __METHOD__);
                     throw $exception;
@@ -74,15 +53,8 @@ class SaveBulkEditJob extends BaseJob
                     throw $throwable;
                 }
 
-                if (($currentRow + 1) === (int)$totalSteps) {
-                    try {
-                        $this->context->delete();
-                    } catch (\Exception $exception) {
-                        throw new Exception('Couldn’t delete context: ' . $exception->getMessage());
-                    }
-                }
-
-                $this->setProgress($queue, ($currentRow + 1) / $totalSteps, 'Element ' . ($currentRow + 1) . ' of ' . $totalSteps);
+                $this->setProgress($queue, ($currentRow + 1) / $totalSteps,
+                    'Element ' . ($currentRow + 1) . ' of ' . $totalSteps);
                 ++$currentRow;
             }
         } catch (\Exception $exception) {
@@ -96,7 +68,11 @@ class SaveBulkEditJob extends BaseJob
 
     protected function defaultDescription(): ?string
     {
-        $name = $this->context->getOwner()->firstName ?? $this->context->getOwner()->email ?? '';
-        return Craft::t('venveo-bulk-edit', 'Bulk Edit in progress by {name}', ['name' => $name]);
+        if ($this->context->getOwner()) {
+            $name = $this->context->getOwner()->firstName ?? $this->context->getOwner()->email ?? '';
+            return Craft::t('venveo-bulk-edit', 'Bulk Edit in progress by {name}', ['name' => $name]);
+        } else {
+            return Craft::t('venveo-bulk-edit', 'Bulk Edit in progress');
+        }
     }
 }